requirements.txt
streamlit
streamlit_shap
numpy
xgboost
shap
python-dotenv
google-cloud-aiplatform
vertexai
matplotlib
langchain 
langchain-openai 
langchain-google-genai
seaborn
imblearn
scikit-learn
tensorflow